fix error message when fcs-genome align input files do not exist, check filesize after checking file existence
