+++ src/workers/BWAWorker.cpp
@@ -1,4 +1,6 @@
+#include <boost/filesystem.hpp>
 #include <string>
+#include <sys/statvfs.h>
 
 #include "fcs-genome/common.h"
 #include "fcs-genome/config.h"
@@ -44,6 +46,40 @@ void BWAWorker::check() {
   ref_path_ = check_input(ref_path_);
   fq1_path_ = check_input(fq1_path_);
   fq2_path_ = check_input(fq2_path_);
+
+  // check temporary storage
+  auto p = boost::filesystem::path(output_path_);
+  std::string temp_dir = p.parent_path().string();
+
+  struct statvfs diskData;
+  statvfs(temp_dir.c_str(), &diskData);
+  uint64_t available = (diskData.f_bavail * diskData.f_frsize);
+
+  DLOG(INFO) << "The available space in " << temp_dir << " is " << available;
+
+  namespace fs = boost::filesystem;
+  size_t size_fastq = 0;
+
+  std::string ext = fs::extension(fq1_path_);
+  int mult = 1;
+  if (ext == ".gz" ) {
+    mult = 3;
+    size_fastq += 3*fs::file_size(fq1_path_);
+    size_fastq += 3*fs::file_size(fq2_path_);
+  }
+  else {
+    size_fastq += fs::file_size(fq1_path_);
+    size_fastq += fs::file_size(fq2_path_);
+  }
+  DLOG(INFO) << "The size of the input data is " << size_fastq;
+
+  if (available < size_fastq) {
+    LOG(ERROR) << "Not enough space in temporary storage: "
+      << temp_dir << ", "
+      << "the size of the temporary folder should be at least " 
+      << mult << " times the size of input FASTQ files";
+    throw silentExit();
+  }
 }
 
 void BWAWorker::setup() {
